@@ -1,19 +1,20 @@
 module k8gbterratest
 
-go 1.17
+go 1.19
 
 require (
 	github.com/AbsaOSS/gopkg v0.1.3
-	github.com/gruntwork-io/terratest v0.38.6
-	github.com/stretchr/testify v1.7.0
-	gopkg.in/yaml.v2 v2.4.0
+	github.com/gruntwork-io/terratest v0.41.16
+	github.com/stretchr/testify v1.8.1
+	gopkg.in/yaml.v3 v3.0.1
 	k8s.io/api v0.20.6
 	k8s.io/apimachinery v0.20.6
 )
 
 require (
-	cloud.google.com/go v0.83.0 // indirect
-	github.com/aws/aws-sdk-go v1.40.56 // indirect
+	cloud.google.com/go/compute v1.12.1 // indirect
+	cloud.google.com/go/compute/metadata v0.2.1 // indirect
+	github.com/aws/aws-sdk-go v1.44.122 // indirect
 	github.com/boombuler/barcode v1.0.1-0.20190219062509-6c824513bacc // indirect
 	github.com/cpuguy83/go-md2man/v2 v2.0.0 // indirect
 	github.com/davecgh/go-spew v1.1.1 // indirect
