@@ -21,25 +21,24 @@ Generated by GoLic, for more details see: https://github.com/AbsaOSS/golic
 import (
 	"context"
 	"fmt"
-	"io/ioutil"
 	"net"
 	"os"
 	"strings"
 	"testing"
 	"time"
 
-	"github.com/golang/mock/gomock"
-
-	str "github.com/AbsaOSS/gopkg/string"
 	k8gbv1beta1 "github.com/k8gb-io/k8gb/api/v1beta1"
 	"github.com/k8gb-io/k8gb/controllers/depresolver"
 	"github.com/k8gb-io/k8gb/controllers/internal/utils"
 	"github.com/k8gb-io/k8gb/controllers/logging"
+	"github.com/k8gb-io/k8gb/controllers/mocks"
 	"github.com/k8gb-io/k8gb/controllers/providers/assistant"
 	"github.com/k8gb-io/k8gb/controllers/providers/dns"
 	"github.com/k8gb-io/k8gb/controllers/providers/metrics"
 	"github.com/k8gb-io/k8gb/controllers/tracing"
 
+	str "github.com/AbsaOSS/gopkg/string"
+	"github.com/golang/mock/gomock"
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/prometheus/client_golang/prometheus/testutil"
 	"github.com/stretchr/testify/assert"
@@ -296,7 +295,7 @@ func TestIngressHostsPerStatusMetricReflectionForNotFoundStatus(t *testing.T) {
 func TestHealthyRecordMetric(t *testing.T) {
 	// arrange
 	expectedHealthyRecordsMetricCount := 3.0
-	ingressIPs := []corev1.LoadBalancerIngress{
+	ingressIPs := []netv1.IngressLoadBalancerIngress{
 		{IP: "10.0.0.1"},
 		{IP: "10.0.0.2"},
 		{IP: "10.0.0.3"},
@@ -340,7 +339,7 @@ func TestGslbErrorsIncrement(t *testing.T) {
 	defer ctrl.Finish()
 	settings := provideSettings(t, predefinedConfig)
 	var label = prometheus.Labels{"namespace": settings.gslb.Namespace, "name": settings.gslb.Name}
-	m := dns.NewMockProvider(ctrl)
+	m := mocks.NewMockProvider(ctrl)
 	cnt := testutil.ToFloat64(metrics.Metrics().Get(metrics.K8gbGslbErrorsTotal).AsCounterVec().With(label))
 	m.EXPECT().GslbIngressExposedIPs(gomock.Any()).Return([]string{}, nil).Times(1)
 	m.EXPECT().SaveDNSEndpoint(gomock.Any(), gomock.Any()).Return(fmt.Errorf("save DNS error")).Times(1)
@@ -377,7 +376,7 @@ func TestGslbCreatesDNSEndpointCRForHealthyIngressHosts(t *testing.T) {
 			Targets:    externaldns.Targets{"10.0.0.1", "10.0.0.2", "10.0.0.3"},
 			Labels:     externaldns.Labels{"strategy": "roundRobin"}},
 	}
-	ingressIPs := []corev1.LoadBalancerIngress{
+	ingressIPs := []netv1.IngressLoadBalancerIngress{
 		{IP: "10.0.0.1"},
 		{IP: "10.0.0.2"},
 		{IP: "10.0.0.3"},
@@ -408,7 +407,7 @@ func TestDNSRecordReflectionInStatus(t *testing.T) {
 	serviceName := defaultPodinfoServiceName
 	dnsEndpoint := &externaldns.DNSEndpoint{}
 	want := map[string][]string{"roundrobin.cloud.example.com": {"10.0.0.1", "10.0.0.2", "10.0.0.3"}}
-	ingressIPs := []corev1.LoadBalancerIngress{
+	ingressIPs := []netv1.IngressLoadBalancerIngress{
 		{IP: "10.0.0.1"},
 		{IP: "10.0.0.2"},
 		{IP: "10.0.0.3"},
@@ -437,7 +436,7 @@ func TestLocalDNSRecordsHasSpecialAnnotation(t *testing.T) {
 	serviceName := defaultPodinfoServiceName
 	dnsEndpoint := &externaldns.DNSEndpoint{}
 	want := "local"
-	ingressIPs := []corev1.LoadBalancerIngress{
+	ingressIPs := []netv1.IngressLoadBalancerIngress{
 		{IP: "10.0.0.1"},
 		{IP: "10.0.0.2"},
 		{IP: "10.0.0.3"},
@@ -489,7 +488,7 @@ func TestCanGetExternalTargetsFromK8gbInAnotherLocation(t *testing.T) {
 			Labels:     externaldns.Labels{"strategy": "roundRobin"}},
 	}
 	hrWant := map[string][]string{"roundrobin.cloud.example.com": {"10.0.0.1", "10.0.0.2", "10.0.0.3", "10.1.0.1", "10.1.0.2", "10.1.0.3"}}
-	ingressIPs := []corev1.LoadBalancerIngress{
+	ingressIPs := []netv1.IngressLoadBalancerIngress{
 		{IP: "10.0.0.1"},
 		{IP: "10.0.0.2"},
 		{IP: "10.0.0.3"},
@@ -573,7 +572,7 @@ func TestReturnsOwnRecordsUsingFailoverStrategyWhenPrimary(t *testing.T) {
 			Labels:     externaldns.Labels{"strategy": depresolver.FailoverStrategy},
 		},
 	}
-	ingressIPs := []corev1.LoadBalancerIngress{
+	ingressIPs := []netv1.IngressLoadBalancerIngress{
 		{IP: "10.0.0.1"},
 		{IP: "10.0.0.2"},
 		{IP: "10.0.0.3"},
@@ -628,7 +627,7 @@ func TestReturnsExternalRecordsUsingFailoverStrategy(t *testing.T) {
 			Labels:     externaldns.Labels{"strategy": depresolver.FailoverStrategy},
 		},
 	}
-	ingressIPs := []corev1.LoadBalancerIngress{
+	ingressIPs := []netv1.IngressLoadBalancerIngress{
 		{IP: "10.0.0.1"},
 		{IP: "10.0.0.2"},
 		{IP: "10.0.0.3"},
@@ -696,7 +695,7 @@ func TestReturnsExternalRecordsUsingFailoverStrategyAndFallbackDNSserver(t *test
 			Labels:     externaldns.Labels{"strategy": depresolver.FailoverStrategy},
 		},
 	}
-	ingressIPs := []corev1.LoadBalancerIngress{
+	ingressIPs := []netv1.IngressLoadBalancerIngress{
 		{IP: "10.0.0.1"},
 		{IP: "10.0.0.2"},
 	}
@@ -984,7 +983,7 @@ func TestResolvesLoadBalancerHostnameFromIngressStatus(t *testing.T) {
 	err := settings.client.Get(context.TODO(), settings.request.NamespacedName, settings.ingress)
 	require.NoError(t, err, "Failed to get expected ingress")
 
-	settings.ingress.Status.LoadBalancer.Ingress = []corev1.LoadBalancerIngress{{Hostname: "one.one.one.one"}}
+	settings.ingress.Status.LoadBalancer.Ingress = []netv1.IngressLoadBalancerIngress{{Hostname: "one.one.one.one"}}
 	err = settings.client.Status().Update(context.TODO(), settings.ingress)
 	require.NoError(t, err, "Failed to update gslb Ingress Address")
 
@@ -1280,7 +1279,7 @@ func provideSettings(t *testing.T, expected depresolver.Config) (settings testSe
 	if os.IsNotExist(err) {
 		t.Fatalf("Sample CR yaml file not found at: %s", crSampleYaml)
 	}
-	gslbYaml, err := ioutil.ReadFile(crSampleYaml)
+	gslbYaml, err := os.ReadFile(crSampleYaml)
 	if err != nil {
 		t.Fatalf("Can't open example CR file: %s", crSampleYaml)
 	}
