@@ -19,9 +19,6 @@ Generated by GoLic, for more details see: https://github.com/AbsaOSS/golic
 */
 
 import (
-	"strconv"
-	"strings"
-
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
 
@@ -33,8 +30,8 @@ import (
 type Strategy struct {
 	// Load balancing strategy type:(roundRobin|failover)
 	Type string `json:"type"`
-	// roundrobin and in the future consistent may (but also may not) contain a weight
-	Weight Weight `json:"weight,omitempty"`
+	// Weight is defined by map region:weight
+	Weight map[string]int `json:"weight,omitempty"`
 	// Primary Geo Tag. Valid for failover strategy only
 	PrimaryGeoTag string `json:"primaryGeoTag,omitempty"`
 	// Defines DNS record TTL in seconds
@@ -102,29 +99,4 @@ func (h HealthStatus) String() string {
 
 func init() {
 	SchemeBuilder.Register(&Gslb{}, &GslbList{})
-}
-
-type Percentage string
-
-type Weight map[string]Percentage
-
-func (p Percentage) String() string {
-	return string(p)
-}
-
-func (p Percentage) IsEmpty() bool {
-	return string(p) == ""
-}
-
-func (p Percentage) TryParse() (v int, err error) {
-	return strconv.Atoi(strings.TrimSuffix(strings.ReplaceAll(p.String(), " ", ""), "%"))
-}
-
-func (p Percentage) Int() int {
-	v, _ := p.TryParse()
-	return v
-}
-
-func (w Weight) IsEmpty() bool {
-	return len(w) == 0
 }
