@@ -175,7 +175,7 @@ func (in *Strategy) DeepCopyInto(out *Strategy) {
 	*out = *in
 	if in.Weight != nil {
 		in, out := &in.Weight, &out.Weight
-		*out = make(Weight, len(*in))
+		*out = make(map[string]int, len(*in))
 		for key, val := range *in {
 			(*out)[key] = val
 		}
@@ -190,25 +190,4 @@ func (in *Strategy) DeepCopy() *Strategy {
 	out := new(Strategy)
 	in.DeepCopyInto(out)
 	return out
-}
-
-// DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
-func (in Weight) DeepCopyInto(out *Weight) {
-	{
-		in := &in
-		*out = make(Weight, len(*in))
-		for key, val := range *in {
-			(*out)[key] = val
-		}
-	}
-}
-
-// DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new Weight.
-func (in Weight) DeepCopy() Weight {
-	if in == nil {
-		return nil
-	}
-	out := new(Weight)
-	in.DeepCopyInto(out)
-	return *out
 }
