@@ -69,29 +69,19 @@ func (dr *DependencyResolver) validateSpec(strategy k8gbv1beta1.Strategy) (err e
 	if err != nil {
 		return
 	}
-	if !strategy.Weight.IsEmpty() {
+	if len(strategy.Weight) != 0 {
 		if strategy.Type != RoundRobinStrategy {
 			return fmt.Errorf(`weight is allowed only for roundRobin strategy`)
 		}
-		summary := 0
 		for k, w := range strategy.Weight {
-			var wi int
 			err = field("weight", k).isNotEmpty().matchRegexp(geoTagRegex).err
 			if err != nil {
 				return err
 			}
-			wi, err = w.TryParse()
-			if err != nil {
-				return fmt.Errorf("weight invalid value %s", w.String())
-			}
-			err = field("weight", wi).isHigherOrEqualToZero().isLessOrEqualTo(100).err
+			err = field("weight", w).isHigherOrEqualToZero().isLessOrEqualTo(1000).err
 			if err != nil {
 				return err
 			}
-			summary += wi
-		}
-		if summary != 100 {
-			return fmt.Errorf("weight the sum of the weights must be equal to 100")
 		}
 	}
 	return nil
