@@ -5,11 +5,12 @@ import (
 	"fmt"
 	"testing"
 
-	"github.com/golang/mock/gomock"
 	k8gbv1beta1 "github.com/k8gb-io/k8gb/api/v1beta1"
 	"github.com/k8gb-io/k8gb/controllers/depresolver"
+	"github.com/k8gb-io/k8gb/controllers/mocks"
 	"github.com/k8gb-io/k8gb/controllers/providers/assistant"
-	"github.com/k8gb-io/k8gb/controllers/providers/dns"
+
+	"github.com/golang/mock/gomock"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"sigs.k8s.io/controller-runtime/pkg/client"
@@ -37,7 +38,7 @@ Generated by GoLic, for more details see: https://github.com/AbsaOSS/golic
 func TestWeight(t *testing.T) {
 	// arrange
 	type wrr struct {
-		weight  string
+		weight  int
 		targets []string
 	}
 	var tests = []struct {
@@ -50,9 +51,9 @@ func TestWeight(t *testing.T) {
 			name:          "eu35-us50-za15",
 			injectWeights: true,
 			data: map[string]wrr{
-				"eu": {weight: "35%", targets: []string{"10.10.0.1", "10.10.0.2"}},
-				"us": {weight: "50%", targets: []string{"10.0.0.1", "10.0.0.2"}},
-				"za": {weight: "15%", targets: []string{"10.22.0.1", "10.22.0.2", "10.22.1.1"}},
+				"eu": {weight: 35, targets: []string{"10.10.0.1", "10.10.0.2"}},
+				"us": {weight: 50, targets: []string{"10.0.0.1", "10.0.0.2"}},
+				"za": {weight: 15, targets: []string{"10.22.0.1", "10.22.0.2", "10.22.1.1"}},
 			},
 			expectedLabels: map[string]string{
 				"strategy":       depresolver.RoundRobinStrategy,
@@ -69,9 +70,9 @@ func TestWeight(t *testing.T) {
 			name:          "eu100-us0-za0",
 			injectWeights: true,
 			data: map[string]wrr{
-				"eu": {weight: "100%", targets: []string{"10.10.0.1", "10.10.0.2"}},
-				"us": {weight: "0%", targets: []string{"10.0.0.1", "10.0.0.2"}},
-				"za": {weight: "0%", targets: []string{"10.22.0.1", "10.22.0.2", "10.22.1.1"}},
+				"eu": {weight: 100, targets: []string{"10.10.0.1", "10.10.0.2"}},
+				"us": {weight: 0, targets: []string{"10.0.0.1", "10.0.0.2"}},
+				"za": {weight: 0, targets: []string{"10.22.0.1", "10.22.0.2", "10.22.1.1"}},
 			},
 			expectedLabels: map[string]string{
 				"strategy":        depresolver.RoundRobinStrategy,
@@ -88,9 +89,9 @@ func TestWeight(t *testing.T) {
 			name:          "weights-without-external-targets",
 			injectWeights: true,
 			data: map[string]wrr{
-				"eu": {weight: "25%", targets: []string{}},
-				"us": {weight: "75%", targets: []string{}},
-				"za": {weight: "0%", targets: []string{}},
+				"eu": {weight: 25, targets: []string{}},
+				"us": {weight: 75, targets: []string{}},
+				"za": {weight: 0, targets: []string{}},
 			},
 			expectedLabels: map[string]string{
 				"strategy": depresolver.RoundRobinStrategy,
@@ -100,9 +101,9 @@ func TestWeight(t *testing.T) {
 			name:          "weights-without-external-targets",
 			injectWeights: true,
 			data: map[string]wrr{
-				"eu": {weight: "25%", targets: []string{}},
-				"us": {weight: "75%", targets: []string{}},
-				"za": {weight: "0%", targets: []string{}},
+				"eu": {weight: 25, targets: []string{}},
+				"us": {weight: 75, targets: []string{}},
+				"za": {weight: 0, targets: []string{}},
 			},
 			expectedLabels: map[string]string{
 				"strategy": depresolver.RoundRobinStrategy,
@@ -120,9 +121,9 @@ func TestWeight(t *testing.T) {
 			name:          "no weights with external targets",
 			injectWeights: false,
 			data: map[string]wrr{
-				"eu": {weight: "100%", targets: []string{"10.10.0.1", "10.10.0.2"}},
-				"us": {weight: "0%", targets: []string{"10.0.0.1", "10.0.0.2"}},
-				"za": {weight: "0%", targets: []string{"10.22.0.1", "10.22.0.2", "10.22.1.1"}},
+				"eu": {weight: 100, targets: []string{"10.10.0.1", "10.10.0.2"}},
+				"us": {weight: 0, targets: []string{"10.0.0.1", "10.0.0.2"}},
+				"za": {weight: 0, targets: []string{"10.22.0.1", "10.22.0.2", "10.22.1.1"}},
 			},
 			expectedLabels: map[string]string{
 				"strategy": depresolver.RoundRobinStrategy,
@@ -146,9 +147,9 @@ func TestWeight(t *testing.T) {
 				if !test.injectWeights {
 					return nil
 				}
-				gslb.Spec.Strategy.Weight = make(map[string]k8gbv1beta1.Percentage, 0)
+				gslb.Spec.Strategy.Weight = make(map[string]int, 0)
 				for k, w := range test.data {
-					gslb.Spec.Strategy.Weight[k] = k8gbv1beta1.Percentage(w.weight)
+					gslb.Spec.Strategy.Weight[k] = w.weight
 				}
 				return nil
 			}
@@ -169,8 +170,8 @@ func TestWeight(t *testing.T) {
 			ctrl := gomock.NewController(t)
 			defer ctrl.Finish()
 			// settings := provideSettings(t, predefinedConfig)
-			m := dns.NewMockProvider(ctrl)
-			r := depresolver.NewMockResolver(ctrl)
+			m := mocks.NewMockProvider(ctrl)
+			r := mocks.NewMockGslbResolver(ctrl)
 			m.EXPECT().GslbIngressExposedIPs(gomock.Any()).Return([]string{}, nil).Times(1)
 			m.EXPECT().SaveDNSEndpoint(gomock.Any(), gomock.Any()).Do(assertAnnotation).Return(fmt.Errorf("save DNS error")).Times(1)
 			m.EXPECT().CreateZoneDelegationForExternalDNS(gomock.Any()).Return(nil).AnyTimes()
