@@ -1,8 +1,8 @@
 // Code generated by MockGen. DO NOT EDIT.
 // Source: controllers/providers/assistant/assistant.go
 
-// Package assistant is a generated GoMock package.
-package assistant
+// Package mocks is a generated GoMock package.
+package mocks
 
 /*
 Copyright 2022 The k8gb Contributors.
@@ -28,6 +28,7 @@ import (
 
 	gomock "github.com/golang/mock/gomock"
 	v1beta1 "github.com/k8gb-io/k8gb/api/v1beta1"
+	assistant "github.com/k8gb-io/k8gb/controllers/providers/assistant"
 	endpoint "sigs.k8s.io/external-dns/endpoint"
 )
 
@@ -70,10 +71,10 @@ func (mr *MockAssistantMockRecorder) CoreDNSExposedIPs() *gomock.Call {
 }
 
 // GetExternalTargets mocks base method.
-func (m *MockAssistant) GetExternalTargets(host string, extClusterNsNames map[string]string) Targets {
+func (m *MockAssistant) GetExternalTargets(host string, extClusterNsNames map[string]string) assistant.Targets {
 	m.ctrl.T.Helper()
 	ret := m.ctrl.Call(m, "GetExternalTargets", host, extClusterNsNames)
-	ret0, _ := ret[0].(Targets)
+	ret0, _ := ret[0].(assistant.Targets)
 	return ret0
 }
 
