@@ -1,8 +1,8 @@
 // Code generated by MockGen. DO NOT EDIT.
 // Source: github.com/infobloxopen/infoblox-go-client (interfaces: IBConnector)
 
-// Package dns is a generated GoMock package.
-package dns
+// Package mocks is a generated GoMock package.
+package mocks
 
 /*
 Copyright 2022 The k8gb Contributors.
